Add app's shared database settings to database.yml
This step is necessary for deploying Rails application to Sqale even though your app really doesn't use database storage.

The reason why you need to paste a correct database settings to database.yml is because of Sqale's database environment.
Sqale does not support sqlite database in production. Sqlite is a file-based database, so you cannot enjoy the benefit of containers if you use sqlite database (filesystem storage of each containers are not shared)
For preventing these database devide problem, we decided to drop sqlite support.

When you pushed a Rails app that using sqlite in production, Sqale's build step detect it and automatically drop the sqlite database settings, because build cannot finish if these settings are exists.
Rails app will pass the build step but it may not run on production environment because the producion database settings is not correct. So, you need to paste the correct database settings.